pyproject.toml
[project]
name = "auth-user"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "bcrypt>=5.0.0",
    "fastapi[standard]>=0.136.1",
    "pydantic-settings>=2.14.0",
    "pyjwt>=2.12.1",
]
